# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Todo API",
    "version": "1.0.0",
    "description": "A production-ready REST API for managing todos.\n\n## Features\n- Full CRUD on todos\n- Pagination (`skip` / `limit`) on list\n- Full-text search on title (LIKE)\n- Input validation via Zod (structured error arrays)\n- SQLite persistence via sql.js\n\n## Status values\n| Value | Meaning |\n|-------|---------|\n| `pending` | Not started yet (default) |\n| `in-progress` | Currently being worked on |\n| `done` | Completed |",
    "contact": {
      "name": "lootopia-les-3-dev",
      "url": "https://github.com/lootopia-les-3-dev/todo-api-node"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "schemas": {
      "Todo": {
        "type": "object",
        "description": "Un todo stocké en base de données",
        "required": [
          "id",
          "title",
          "status"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identifiant unique auto-incrémenté, généré par la BDD",
            "example": 42,
            "readOnly": true
          },
          "title": {
            "type": "string",
            "description": "Titre court de la tâche (1–200 caractères)",
            "minLength": 1,
            "maxLength": 200,
            "example": "Buy groceries"
          },
          "description": {
            "type": "string",
            "description": "Description longue optionnelle (max 1000 caractères). Peut être null.",
            "maxLength": 1000,
            "nullable": true,
            "example": "Oat milk, sourdough bread, eggs"
          },
          "status": {
            "type": "string",
            "description": "Statut actuel du cycle de vie du todo",
            "enum": [
              "pending",
              "in-progress",
              "done"
            ],
            "default": "pending",
            "example": "in-progress"
          }
        },
        "example": {
          "id": 42,
          "title": "Buy groceries",
          "description": "Oat milk, sourdough bread, eggs",
          "status": "in-progress"
        }
      },
...
      "ValidationError": {
        "type": "object",
        "description": "Retourné quand le body ou les query params échouent la validation Zod.",
        "properties": {
          "detail": {
            "description": "Tableau d'issues Zod — chaque issue a un `path` et un `message`",
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": [
                        "title"
                      ]
                    },
                    "message": {
                      "type": "string",
                      "example": "String must contain at least 1 character(s)"
                    }
                  }
                }
              },
              {
                "type": "string"
              }
            ],
            "example": [
              {
                "path": [
                  "title"
                ],
                "message": "Required"
              },
              {
                "path": [
                  "status"
                ],
                "message": "Invalid enum value. Expected 'pending' | 'in-progress' | 'done'"
              }
            ]
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/todos/search": {
      "get": {
        "tags": [
          "Todos"
        ],
        "summary": "Rechercher des todos par titre",
        "description": "Effectue une recherche **insensible à la casse** sur le champ `title` avec une correspondance partielle (LIKE `%q%`).\n\nRetourne un tableau vide si aucun résultat — jamais un 404.\n\n**Exemples de recherche :**\n- `?q=buy` → trouve \"Buy groceries\", \"Buy milk\"\n- `?q=CALL` → trouve \"Call the dentist\" (insensible à la casse)\n- `?q=zzz` → `[]` si aucun match\n",
        "operationId": "searchTodos",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "description": "Terme recherché dans les titres (1–200 caractères, insensible à la casse)",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200,
              "example": "grocery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des todos dont le titre contient le terme recherché",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Todo"
                  }
                },
                "examples": {
                  "plusieursResultats": {
                    "summary": "Plusieurs résultats trouvés",
                    "value": [
                      {
                        "id": 1,
                        "title": "Buy groceries",
                        "description": "Oat milk, eggs",
                        "status": "pending"
                      },
                      {
                        "id": 5,
                        "title": "Buy milk",
                        "description": null,
                        "status": "done"
                      }
                    ]
                  },
                  "unResultat": {
                    "summary": "Un seul résultat",
                    "value": [
                      {
                        "id": 3,
                        "title": "Call the dentist",
                        "description": "Tuesday at 10am",
                        "status": "in-progress"
                      }
                    ]
                  },
                  "aucunResultat": {
                    "summary": "Aucun résultat (terme introuvable)",
                    "value": []
                  }
                }
              }
            }
          },
          "400": {
            "description": "Paramètre `q` manquant ou trop court/long",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                },
                "examples": {
                  "manquant": {
                    "summary": "Paramètre q absent",
                    "value": {
                      "detail": [
                        {
                          "path": [
                            "q"
                          ],
                          "message": "Required"
                        }
                      ]
                    }
                  },
                  "tropCourt": {
                    "summary": "Paramètre q vide",
                    "value": {
                      "detail": [
                        {
                          "path": [
                            "q"
                          ],
                          "message": "String must contain at least 1 character(s)"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}